requirements.txt
flask>=2.3,<4
flask-cors>=4.0,<5
openpyxl>=3.1,<4
requests>=2.28,<3
gspread>=6.0,<7
google-auth>=2.0,<3
google-auth-oauthlib>=1.0,<2
google-api-python-client>=2.0,<3
